requirements.txt
librosa==0.10.1
openai-whisper==20231117
resemblyzer==0.1.1
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
spacy>=3.7.0
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.0/en_core_web_lg-3.7.0-py3-none-any.whl
scikit-learn>=1.3.0
numpy>=1.24.0
nltk>=3.8.0
streamlit>=1.28.0
google-generativeai>=0.3.0
fpdf>=1.7.0
soundfile>=0.12.0